+++ ProyectoFinal/src/visual/ListaProveedores.java
@@ -149,12 +149,18 @@ public class ListaProveedores extends JDialog {
         int selectedRow = table.getSelectedRow();
         if (selectedRow >= 0) {
             String idProveedor = (String) tableModel.getValueAt(selectedRow, 0);
-            Tienda.getInstance().eliminarPersona(idProveedor);
-            tableModel.removeRow(selectedRow);
-            JOptionPane.showMessageDialog(this, "Proveedor eliminado correctamente.");
+            int confirmacion = JOptionPane.showConfirmDialog(this,"¿Estas seguro de que deseas eliminar este proveedor?","Confirmacion de eliminacion",JOptionPane.YES_NO_OPTION);
+            if (confirmacion == JOptionPane.YES_OPTION) {
+                Tienda.getInstance().eliminarPersona(idProveedor);
+                tableModel.removeRow(selectedRow);
+                JOptionPane.showMessageDialog(this, "Proveedor eliminado correctamente.");
+            } else {
+                JOptionPane.showMessageDialog(this, "Eliminacion cancelada.");
+            }
         } else {
             JOptionPane.showMessageDialog(this, "Por favor, seleccione un proveedor para eliminar.");
         }
     }
+
 
 }
